refactor(stelae): make the chained-publish lifecycle profile-generic
`stelae-driver` gains `DriverProfile` — four questions the publish and
restore lifecycle has to ask a profile that `stelae::Profile` deliberately
does not answer — and with it the `Predecessor` seam and the whole
chained-publish lifecycle that used to live in `crates/snapshot`.

Moved, generic over `&dyn DriverProfile` and unchanged in what they
canonicalize: `Predecessor` and `First`; `Origin`, `PublishRecord` and its
`Recording`; `Chained` and `inheritable_layers`; `Publishing`, `Tuning`,
`open` and `standing`. Every document that crosses the boundary is opaque
`serde_json::Value` composed profile-side, and the driver never derives a
node's layout — `record_path_in` stays in `crates/snapshot` and hands the
resumption record's path in.

Refusals keep the variant a caller matches and the message an operator
reads: the driver's new `DatasetMismatch` flattens back into this profile's
`NetworkMismatch`, on the precedent the other shared refusals already set.

`Publishing::storage_path` is now `record_path: Option<PathBuf>`, and the
type is `Clone` rather than `Copy` because of it: a builder taking a `&Path`
would accept a node's storage directory as readily as its record, silently,
and the two are one `join` apart.
