+++ crates/stelae-driver/Cargo.toml
@@ -20,6 +20,7 @@ stelae = { path = "../stelae" }
 
 fs4.workspace = true
 minicbor.workspace = true
+serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
